+++ gtsfm/multi_view_optimizer.py
@@ -116,35 +116,6 @@ class MultiViewOptimizer:
         return ba_input_graph, ba_result_graph, saved_metrics_graph, react_saved_metrics_graph
 
 
-def prune_to_largest_connected_component(
-    rotations: Dict[Tuple[int, int], Optional[Rot3]], unit_translations: Dict[Tuple[int, int], Optional[Unit3]],
-) -> Tuple[Dict[Tuple[int, int], Rot3], Dict[Tuple[int, int], Unit3]]:
-    """Process the graph of image indices with Rot3s/Unit3s defining edges, and select the largest connected component.
-
-    Args:
-        rotations: dictionary of relative rotations for pairs.
-        unit_translations: dictionary of relative unit-translations for pairs.
-
-    Returns:
-        Subset of rotations which are in the largest connected components.
-        Subset of unit_translations which are in the largest connected components.
-    """
-    input_edges = [k for (k, v) in rotations.items() if v is not None]
-    nodes_in_pruned_graph = graph_utils.get_nodes_in_largest_connected_component(input_edges)
-
-    # select the edges with nodes in the pruned graph
-    selected_edges = []
-    for i1, i2 in rotations.keys():
-        if i1 in nodes_in_pruned_graph and i2 in nodes_in_pruned_graph:
-            selected_edges.append((i1, i2))
-
-    # return the subset of original input
-    return (
-        {k: rotations[k] for k in selected_edges},
-        {k: unit_translations[k] for k in selected_edges},
-    )
-
-
 def init_cameras(
     wRi_list: List[Optional[Rot3]], wti_list: List[Optional[Point3]], intrinsics_list: List[Cal3Bundler],
 ) -> Dict[int, PinholeCameraCal3Bundler]:
@@ -154,6 +125,7 @@ def init_cameras(
         wRi_list: rotations for cameras.
         wti_list: translations for cameras.
         intrinsics_list: intrinsics for cameras.
+
     Returns:
         Valid cameras.
     """
